[FIX] runbot_gitlab: Retrieve correctly SSH keys when under GitLab CI (#177)
Currently, SSH keys are not retrieved correctly, due to a bad parsing of
usernames when retrieving commit information. The username is not a
string, but a dictionary `{'login': <user_name>}`.

This commit parses correctly usernames of committers and authors, so SSH
keys are retrieved correctly and SSH access may be granted.

This also causes empty authors to be skipped, they of course have
neither username nor SSH keys.